requirements.txt
# Telegram Bot
python-telegram-bot>=20.7

# Google Gemini AI
google-generativeai>=0.3.0

# Environment Variables
python-dotenv>=1.0.0

# Timezone Support
pytz>=2023.3

# Async Support
asyncio>=3.4.3

# Logging
colorlog>=6.7.0

# Optional: Database Support
# sqlalchemy>=2.0.0
# aiosqlite>=0.19.0

# Optional: Redis for caching
# redis>=5.0.0
# aioredis>=2.0.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0